refactor: extract storage implementation to a separate file for reuse in multiple graphs
